Extract interface from VirtusizeAPIService for proper unit-testing
