+++ virtusize/src/main/java/com/virtusize/android/network/VirtusizeAPIService.kt
@@ -1,30 +1,12 @@
 package com.virtusize.android.network
 
 import android.content.Context
-import android.content.res.Configuration
 import android.graphics.Bitmap
-import android.graphics.BitmapFactory
-import android.view.WindowManager
-import com.virtusize.android.BuildConfig
-import com.virtusize.android.R
-import com.virtusize.android.SharedPreferencesHelper
-import com.virtusize.android.data.local.StoreId
-import com.virtusize.android.data.local.VirtusizeErrorType
 import com.virtusize.android.data.local.VirtusizeEvent
 import com.virtusize.android.data.local.VirtusizeLanguage
 import com.virtusize.android.data.local.VirtusizeMessageHandler
 import com.virtusize.android.data.local.VirtusizeOrder
 import com.virtusize.android.data.local.VirtusizeProduct
-import com.virtusize.android.data.local.virtusizeError
-import com.virtusize.android.data.parsers.BodyProfileRecommendedSizeJsonParser
-import com.virtusize.android.data.parsers.ProductCheckDataJsonParser
-import com.virtusize.android.data.parsers.ProductMetaDataHintsJsonParser
-import com.virtusize.android.data.parsers.ProductTypeJsonParser
-import com.virtusize.android.data.parsers.StoreJsonParser
-import com.virtusize.android.data.parsers.StoreProductJsonParser
-import com.virtusize.android.data.parsers.UserBodyProfileJsonParser
-import com.virtusize.android.data.parsers.UserProductJsonParser
-import com.virtusize.android.data.parsers.UserSessionInfoJsonParser
 import com.virtusize.android.data.remote.BodyProfileRecommendedSize
 import com.virtusize.android.data.remote.Product
 import com.virtusize.android.data.remote.ProductCheckData
@@ -34,22 +16,10 @@ import com.virtusize.android.data.remote.Store
 import com.virtusize.android.data.remote.UserBodyProfile
 import com.virtusize.android.data.remote.UserSessionInfo
 import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
 import org.json.JSONObject
-import java.net.URL
-import java.util.Locale
 import javax.net.ssl.HttpsURLConnection
 
-/**
- * A class that handles API requests to the Virtusize server
- * @param context the application context
- * @param messageHandler pass VirtusizeMessageHandler to listen to any Virtusize-related messages
- */
-internal open class VirtusizeAPIService(
-    private val context: Context,
-    private val messageHandler: VirtusizeMessageHandler?,
-) {
+internal interface VirtusizeAPIService {
     companion object {
         private var instance: VirtusizeAPIService? = null
 
@@ -61,92 +31,41 @@ internal open class VirtusizeAPIService(
             messageHandler: VirtusizeMessageHandler?,
         ): VirtusizeAPIService {
             if (instance == null) {
-                synchronized(VirtusizeAPIService::javaClass) {
-                    instance = VirtusizeAPIService(context, messageHandler)
+                synchronized(VirtusizeAPIServiceImpl::javaClass) {
+                    instance = VirtusizeAPIServiceImpl(context, messageHandler)
                 }
             }
             return instance!!
         }
     }
 
-    // The helper to store data locally using Shared Preferences
-    private var sharedPreferencesHelper: SharedPreferencesHelper =
-        SharedPreferencesHelper.getInstance(context)
-
-    // Device screen resolution
-    private lateinit var resolution: String
-
-    // The HTTP URL connection that is used to make a single request
-    private var httpURLConnection: HttpsURLConnection? = null
-
-    // The dispatcher that determines what thread the corresponding coroutine uses for its execution
-    private var coroutineDispatcher: CoroutineDispatcher = Dispatchers.IO
-
     /**
      * Sets the HTTP URL connection
      * @param urlConnection an instance of [HttpsURLConnection]
      */
-    internal fun setHTTPURLConnection(urlConnection: HttpsURLConnection?) {
-        this.httpURLConnection = urlConnection
-    }
+    fun setHTTPURLConnection(urlConnection: HttpsURLConnection?)
 
     /**
      * Sets the Coroutine dispatcher
      * @param dispatcher an instance of [CoroutineDispatcher]
      */
-    internal fun setCoroutineDispatcher(dispatcher: CoroutineDispatcher) {
-        this.coroutineDispatcher = dispatcher
-    }
+    fun setCoroutineDispatcher(dispatcher: CoroutineDispatcher)
 
-    internal suspend fun fetchLatestAoyamaVersion(): VirtusizeApiResponse<String> =
-        withContext(Dispatchers.IO) {
-            val apiRequest = VirtusizeApi.fetchLatestAoyamaVersion()
-            VirtusizeApiTask(
-                httpURLConnection,
-                sharedPreferencesHelper,
-                messageHandler,
-            ).setResponseFormat(VirtusizeApiResponseFormat.STRING).execute(apiRequest)
-        }
+    suspend fun fetchLatestAoyamaVersion(): VirtusizeApiResponse<String>
 
     /**
      * Executes the API task to make a network request for Product Check
      * @param product [VirtusizeProduct]
      * @return the [VirtusizeApiResponse] with the data class [ProductCheckData]
      */
-    internal suspend fun productCheck(product: VirtusizeProduct): VirtusizeApiResponse<ProductCheckData> =
-        withContext(Dispatchers.IO) {
-            val apiRequest = VirtusizeApi.productCheck(product)
-            VirtusizeApiTask(
-                httpURLConnection,
-                sharedPreferencesHelper,
-                messageHandler,
-            )
-                .setJsonParser(ProductCheckDataJsonParser())
-                .execute<ProductCheckData>(apiRequest)
-                .also { response ->
-                    val storeId = response.successData?.data?.storeId
-                    if (storeId != null) {
-                        VirtusizeApi.setStoreId(StoreId(storeId))
-                    }
-                }
-        }
+    suspend fun productCheck(product: VirtusizeProduct): VirtusizeApiResponse<ProductCheckData>
 
     /**
      * Sends an image URL of VirtusizeProduct to the Virtusize server
      * @param product [VirtusizeProduct]
      * @return the [VirtusizeApiResponse] with the data class [ProductMetaDataHints]
      */
-    internal suspend fun sendProductImageToBackend(product: VirtusizeProduct): VirtusizeApiResponse<ProductMetaDataHints> =
-        withContext(Dispatchers.IO) {
-            val apiRequest = VirtusizeApi.sendProductImageToBackend(product = product)
-            VirtusizeApiTask(
-                httpURLConnection,
-                sharedPreferencesHelper,
-                messageHandler,
-            )
-                .setJsonParser(ProductMetaDataHintsJsonParser())
-                .execute(apiRequest)
-        }
+    suspend fun sendProductImageToBackend(product: VirtusizeProduct): VirtusizeApiResponse<ProductMetaDataHints>
 
     /**
      * Sends an event to the Virtusize server
@@ -154,183 +73,62 @@ internal open class VirtusizeAPIService(
      * @param withDataProduct ProductCheckResponse corresponding to VirtusizeProduct
      * @return the [VirtusizeApiResponse]
      */
-    internal suspend fun sendEvent(
+    suspend fun sendEvent(
         event: VirtusizeEvent,
         withDataProduct: ProductCheckData? = null,
-    ): VirtusizeApiResponse<Any> =
-        withContext(Dispatchers.IO) {
-            val defaultDisplay =
-                (context.getSystemService(Context.WINDOW_SERVICE) as WindowManager).defaultDisplay
-            resolution = "${defaultDisplay.height}x${defaultDisplay.width}"
-
-            val apiRequest =
-                VirtusizeApi.sendEventToAPI(
-                    virtusizeEvent = event,
-                    productCheckData = withDataProduct,
-                    deviceOrientation =
-                        if (context.resources.configuration.orientation == Configuration.ORIENTATION_LANDSCAPE) {
-                            context.getString(R.string.landscape)
-                        } else {
-                            context.getString(R.string.portrait)
-                        },
-                    screenResolution = resolution,
-                    versionName = BuildConfig.VERSION_NANE,
-                )
-            VirtusizeApiTask(
-                httpURLConnection,
-                sharedPreferencesHelper,
-                messageHandler,
-            ).execute(apiRequest)
-        }
+    ): VirtusizeApiResponse<Any>
 
     /**
      * Sends an order to the Virtusize server
      * @return the [VirtusizeApiResponse]
      */
-    internal suspend fun sendOrder(
+    suspend fun sendOrder(
         region: String?,
         order: VirtusizeOrder,
-    ): VirtusizeApiResponse<Any> =
-        withContext(Dispatchers.IO) {
-            // Sets the region from the store info
-            order.setRegion(region)
-            val apiRequest = VirtusizeApi.sendOrder(order)
-            VirtusizeApiTask(
-                httpURLConnection,
-                sharedPreferencesHelper,
-                messageHandler,
-            )
-                .execute(apiRequest)
-        }
+    ): VirtusizeApiResponse<Any>
 
     /**
      * Gets the API response for retrieving the specific store info
      * @return the [VirtusizeApiResponse] with the data class [Store]
      */
-    internal open suspend fun getStoreInfo(): VirtusizeApiResponse<Store> =
-        withContext(Dispatchers.IO) {
-            val apiRequest = VirtusizeApi.getStoreInfo()
-            VirtusizeApiTask(
-                httpURLConnection,
-                sharedPreferencesHelper,
-                messageHandler,
-            )
-                .setJsonParser(StoreJsonParser())
-                .execute(apiRequest)
-        }
+    suspend fun getStoreInfo(): VirtusizeApiResponse<Store>
 
     /**
      * Gets the API response for retrieving the store product info
      * @param productId the ID of the store product
      * @return the [VirtusizeApiResponse] with the data class [Product]
      */
-    internal suspend fun getStoreProduct(productId: Int): VirtusizeApiResponse<Product> =
-        withContext(
-            Dispatchers.IO,
-        ) {
-            if (productId == 0) {
-                return@withContext VirtusizeApiResponse.Error(
-                    VirtusizeErrorType.NullProduct.virtusizeError(),
-                )
-            }
-            val apiRequest = VirtusizeApi.getStoreProductInfo(productId.toString())
-            VirtusizeApiTask(
-                httpURLConnection,
-                sharedPreferencesHelper,
-                messageHandler,
-            )
-                .setJsonParser(StoreProductJsonParser())
-                .execute(apiRequest)
-        }
+    suspend fun getStoreProduct(productId: Int): VirtusizeApiResponse<Product>
 
     /**
      * Gets the API response for retrieving the list of the product types
      * @return the [VirtusizeApiResponse] with a list of [ProductType]
      */
-    internal suspend fun getProductTypes(): VirtusizeApiResponse<List<ProductType>> =
-        withContext(
-            Dispatchers.IO,
-        ) {
-            val apiRequest = VirtusizeApi.getProductTypes()
-            VirtusizeApiTask(
-                httpURLConnection,
-                sharedPreferencesHelper,
-                messageHandler,
-            )
-                .setJsonParser(ProductTypeJsonParser())
-                .execute(apiRequest)
-        }
+    suspend fun getProductTypes(): VirtusizeApiResponse<List<ProductType>>
 
     /**
      * Gets the API response for getting the user session data
      * @return the [VirtusizeApiResponse] with [UserSessionInfo]
      */
-    internal suspend fun getUserSessionInfo(): VirtusizeApiResponse<UserSessionInfo> =
-        withContext(
-            Dispatchers.IO,
-        ) {
-            val apiRequest = VirtusizeApi.getSessions()
-            VirtusizeApiTask(
-                httpURLConnection,
-                sharedPreferencesHelper,
-                messageHandler,
-            )
-                .setJsonParser(UserSessionInfoJsonParser())
-                .execute(apiRequest)
-        }
+    suspend fun getUserSessionInfo(): VirtusizeApiResponse<UserSessionInfo>
 
     /**
      * Gets the API response for deleting a user
      * @return the [VirtusizeApiResponse]
      */
-    internal suspend fun deleteUser(): VirtusizeApiResponse<Any> =
-        withContext(
-            Dispatchers.IO,
-        ) {
-            val apiRequest = VirtusizeApi.deleteUser()
-            VirtusizeApiTask(
-                httpURLConnection,
-                sharedPreferencesHelper,
-                messageHandler,
-            )
-                .execute(apiRequest)
-        }
+    suspend fun deleteUser(): VirtusizeApiResponse<Any>
 
     /**
      * Gets the API response for retrieving a list of user products
      * @return the [VirtusizeApiResponse] with the list of [Product]
      */
-    internal suspend fun getUserProducts(): VirtusizeApiResponse<List<Product>> =
-        withContext(
-            Dispatchers.IO,
-        ) {
-            val apiRequest = VirtusizeApi.getUserProducts()
-            VirtusizeApiTask(
-                httpURLConnection,
-                sharedPreferencesHelper,
-                messageHandler,
-            )
-                .setJsonParser(UserProductJsonParser())
-                .execute(apiRequest)
-        }
+    suspend fun getUserProducts(): VirtusizeApiResponse<List<Product>>
 
     /**
      * Gets the API response for retrieving the current user body profile such as age, height, weight and body measurements
      * @return the [VirtusizeApiResponse] with the data class [UserBodyProfile]
      */
-    internal suspend fun getUserBodyProfile(): VirtusizeApiResponse<UserBodyProfile> =
-        withContext(
-            Dispatchers.IO,
-        ) {
-            val apiRequest = VirtusizeApi.getUserBodyProfile()
-            VirtusizeApiTask(
-                httpURLConnection,
-                sharedPreferencesHelper,
-                messageHandler,
-            )
-                .setJsonParser(UserBodyProfileJsonParser())
-                .execute(apiRequest)
-        }
+    suspend fun getUserBodyProfile(): VirtusizeApiResponse<UserBodyProfile>
 
     /**
      * Gets the API response for retrieving the recommended size based on the user body profile
@@ -339,76 +137,30 @@ internal open class VirtusizeAPIService(
      * @param userBodyProfile the user body profile
      * @return the [VirtusizeApiResponse] with the a list of [BodyProfileRecommendedSize]
      */
-    internal suspend fun getBodyProfileRecommendedSize(
+    suspend fun getBodyProfileRecommendedSize(
         productTypes: List<ProductType>,
         storeProduct: Product,
         userBodyProfile: UserBodyProfile,
-    ): VirtusizeApiResponse<ArrayList<BodyProfileRecommendedSize>?> =
-        withContext(Dispatchers.IO) {
-            val apiRequest = VirtusizeApi.getSize(productTypes, storeProduct, userBodyProfile)
-            VirtusizeApiTask(
-                httpURLConnection,
-                sharedPreferencesHelper,
-                messageHandler,
-            )
-                .setJsonParser(BodyProfileRecommendedSizeJsonParser(storeProduct))
-                .execute(apiRequest)
-        }
-
-    /**
-     * Gets the API response for fetching the i18n localization texts
-     * @param language [VirtusizeLanguage] that is set by a client
-     * @return the [VirtusizeApiResponse] with the JSON data
-     */
-    internal open suspend fun getI18n(language: VirtusizeLanguage?): VirtusizeApiResponse<JSONObject> =
-        withContext(Dispatchers.IO) {
-            val apiRequest =
-                VirtusizeApi.getI18n(
-                    language ?: (
-                        VirtusizeLanguage.entries
-                            .find { it.value == Locale.getDefault().language } ?: VirtusizeLanguage.EN
-                    ),
-                )
-            VirtusizeApiTask(
-                httpURLConnection,
-                sharedPreferencesHelper,
-                messageHandler,
-            )
-                .execute(apiRequest)
-        }
-
-    /**
-     * Gets the API response for fetching the custom i18n texts for specific store
-     * @param storeName the name of the store
-     * @return the [VirtusizeApiResponse] with the JSON data
-     */
-    internal open suspend fun getStoreSpecificI18n(storeName: String): VirtusizeApiResponse<JSONObject> =
-        withContext(Dispatchers.IO) {
-            val apiRequest =
-                VirtusizeApi.getStoreSpecificI18n(storeName)
-            VirtusizeApiTask(
-                httpURLConnection,
-                sharedPreferencesHelper,
-                messageHandler,
-            )
-                .execute(apiRequest)
-        }
+    ): VirtusizeApiResponse<ArrayList<BodyProfileRecommendedSize>?>
 
     /**
      * Loads an image URL and returns the bitmap of the image
      * @param urlString the image URL string
      * @return the Bitmap of the image
      */
-    internal suspend fun loadImage(urlString: String): Bitmap? =
-        withContext(
-            Dispatchers.IO,
-        ) {
-            try {
-                URL(urlString).openStream().use {
-                    return@withContext BitmapFactory.decodeStream(it)
-                }
-            } catch (e: Exception) {
-                return@withContext null
-            }
-        }
+    suspend fun loadImage(urlString: String): Bitmap?
+
+    /**
+     * Gets the API response for fetching the i18n localization texts
+     * @param language [VirtusizeLanguage] that is set by a client
+     * @return the [VirtusizeApiResponse] with the JSON data
+     */
+    suspend fun getI18n(language: VirtusizeLanguage?): VirtusizeApiResponse<JSONObject>
+
+    /**
+     * Gets the API response for fetching the custom i18n texts for specific store
+     * @param storeName the name of the store
+     * @return the [VirtusizeApiResponse] with the JSON data
+     */
+    suspend fun getStoreSpecificI18n(storeName: String): VirtusizeApiResponse<JSONObject>
 }
